+++ core/src/main/java/com/smart/home/es/service/impl/ProductEsServiceImpl.java
@@ -18,7 +18,7 @@ public class ProductEsServiceImpl extends EsQueryService implements ProductEsSer
     private ProductRepository productRepository;
 
     @Override
-    public void deleteById(Integer id) {
+    public void deleteById(Long id) {
         productRepository.deleteById(id);
     }
